pr feedback + changelog
