+++ changelog/v1.18.0-beta28/krt-endpoints.yaml
@@ -1,0 +1,6 @@
+changelog:
+- type: NON_USER_FACING
+  issueLink: https://github.com/solo-io/solo-projects/issues/7036
+  resolvesIssue: true
+  description: >-
+    Add krt handling for k8s endpoints
